+++ Changelog.md
@@ -1,4 +1,6 @@
 Change log:
+* 2022-08-22 - U33.1.1 - 1 new pet from Season 3 Reward Track:
+  * Black Bat
 * 2022-07-25 - U33.1 - 5 new + 3 newly available:
   * Mountaineering Donkey - Reward Track Season 2
   * Friendly Bear Cub, Black Bear Cub, Noble Starling - U33.1